Add a related_status to the dealPhases.list endpoint
This is needed because otherwise integrators don't know which phase is
related to a certain status.

An example: team mobile does a `deals.win` and want to know to which
phase id this will resolve. If they want to do it right now, they have
to refetch the deal using deals.info afterwards.