Don't allow tracking directly on a project for now
All time tracking related code expects trackings on a project to also
have a milestone selected. Since we can't be sure on which milestone a
time tracking belongs while tracking on a contact, it doesn't make sense
to support it.

It's a lot more logical to just require a milestone to be selected if
you time track on a project.