Use web_url instead of canonical_url
Canonical_url looks like the SEO property, which indicates that there
are multiple url's for one page and that that url is the preferred one.
That's why web_url is a more logical name.

We went over some other ideas too:

* app_url => a bit confusing with the mobile app
* html_link => what if we link to a pdf preview? Or a plain text app?
* _url or @url => Couldn't find any other API's with a similar approach

We also thought of putting it in a metadata property, so outside of the
data, but we can't really do that anymore since in a `.list` endpoint,
all returned objects are represented in one `data` array.