Add electronic_invoicing_address to customer endpoints
Note that we put this under invoicing preferences. We'll move payment
terms and preferred_currency also to the invoicing preferences in a BC
breaking change, but let's add electronic_invoicing_address already in
it's final destination instead of putting it in the root and then moving
it later. This will help the Italian integrators to swiftly release the
Italian invoicing integrations.