Make customer a property of intervention address
This way we always refer to an intervention address but allow that
address to be linked to a customer.

Now we don't risk losing the actual work order address when the address
of the customer changes. Also, customers can have multiple addresses
which is not possible to select using the original format.