Provide WorkOrder images using temporary download links
This was concluded from a technical analysis following a concern of the
mobile team that the response of a .info call would be pretty heavy with
multiple full resolution images added to a WorkOrder.

By offering them though a temporary downloads link, they can be lazy
loaded and this will result in quicker feedback times in the app.

Later on we will be able to return multiple variations of an image (i.e.
thumbnails) by having multiple images with the same `id` but different
`variation`.