Sign using the finalize endpoint
We don't want to be able to sign a work order and then still update a
work order. That's why it seems logical to pass a signature while
finalization. This means that finalization will always generate a pdf,
but in one case it will be signed and the other not.