Remove an argument that is never used.
The ended_at time is never used in the implementation.
Since this won't be tackled any time soon we are free to remove it.
This is no breaking change since the request will still work (with the ignored argument) and the behaviour stays the same because it wasn't used before.